% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-data.config.R
\name{data.config}
\alias{data.config}
\title{Configuration of input files}
\usage{
data.config(dataset, train_percent, resampling)
}
\arguments{
\item{dataset}{a dataframe}

\item{train_percent}{numerical value (percentage, in decimals) you want for training data}

\item{resampling}{specific method from caret::trainControl}
}
\value{

}
\description{
Configuration of input files
}
\examples{
# upload the necessary packages
library(mlbench)
library(caret)
# load data
# table available in caret package database
data(PimaIndiansDiabetes)
# rename dataset to keep code below generic
my_df <- PimaIndiansDiabetes
my_df_config <- data.config(my_df, 0.7, "repeatedcv")
my_df_config
}
