% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-component.R
\name{is.PipeComponent}
\alias{is.PipeComponent}
\alias{is.PipeComponentClass}
\title{Is this a PipeComponent class or an object?}
\usage{
is.PipeComponent(x)

is.PipeComponentClass(x)
}
\arguments{
\item{x}{R object}
}
\value{
logical
}
\description{
Checks if x is a PipeComponent class or an object
}
\examples{
is.PipeComponentClass(MLP)    # true
is.PipeComponentClass(mlp())  # false because this is an instance
is.PipeComponent(mlp())       # true
}

