% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-ml.graphic.R
\name{ml.graphic}
\alias{ml.graphic}
\title{Plotting evaluation metrics from different machine learning models}
\usage{
ml.graphic(
  ml_result,
  graphic = c("bwplot", "dotplot", "parallelplot", "splom"),
  ...
)
}
\arguments{
\item{ml_result}{a list of trained machine learning models}

\item{graphic}{user-determined plotting options (bwplot, dotplot, splom or parallelplot)}

\item{...}{your specific graphic params}
}
\value{
graphic
}
\description{
Plotting evaluation metrics from different machine learning models
}
\examples{
# load packages
library(mlbench)
library(caret)
library(mlPipe)
# load data
# table available in caret package database
data(PimaIndiansDiabetes)
# rename dataset to keep code below generic
my_df <- PimaIndiansDiabetes
# now we're going to define our training sample
my_training_sample <- ml.train(my_df, 0.8, seed = 42)
my_training_sample
# and now we're able to train our model
my_models <- ml.trainlist(train_sample = my_training_sample, algorithm = c("rf", "lda", "glm", "knn"), metric = "Accuracy", seed = 42, resampling = "repeatedcv")
my_models
results <- caret::resamples(my_models)
summary(results)
ml.graphic(ml_result = results, graphic = "bwplot")
}
