% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-ml.results.graphic.R
\name{ml.results.graphic}
\alias{ml.results.graphic}
\title{Plotting evaluation metrics from different machine learning models}
\usage{
ml.results.graphic(
  ml_result,
  graphic = c("bwplot", "dotplot", "parallelplot", "splom"),
  ...
)
}
\arguments{
\item{ml_result}{a list of trained machine learning models}

\item{graphic}{user-determined plotting options (bwplot, dotplot, splom or parallelplot)}

\item{...}{your specific graphic params}
}
\value{
graphic
}
\description{
Plotting evaluation metrics from different machine learning models
}
\examples{
# upload the necessary packages
library(mlbench)
library(caret)
library(caretEnsemble)
library(mlPipe)
# load data
# table available in caret package database
data(PimaIndiansDiabetes)
# rename dataset to keep code below generic
my_df <- PimaIndiansDiabetes
# now you're able to train your data
my_models <- ml.trainlist(my_df, 0.8, algorithm = c("rf", "lda", "glm", "knn"), metric = "Accuracy", seed = 42, resampling = "cv")
my_models
results <- caret::resamples(my_models)
summary(results)
ml.results.graphic(ml_result = results, graphic = "bwplot")
}
