% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-ml.testset.R
\name{ml.testset}
\alias{ml.testset}
\title{Setting a test sample to machine learning analyses}
\usage{
ml.testset(dataset, test_percent, seed)
}
\arguments{
\item{dataset}{a dataframe}

\item{test_percent}{numerical value (percentage, in decimals) you want for test data}

\item{seed}{specific seed value}
}
\value{
a new dataframe, now only with our test sample
}
\description{
Setting a test sample to machine learning analyses
}
\examples{
# load packages
library(mlbench)
library(caret)
library(mlPipe)
# load data
# table available in caret package database
data(PimaIndiansDiabetes)
my_df <- PimaIndiansDiabetes
# rename dataset to keep code below generic
my_df <- PimaIndiansDiabetes
# now you're able to define your training sample
my_test_sample <- ml.testset(my_df, 0.2, seed = 42)
my_test_sample
}
