% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-ml.train.R
\name{ml.train}
\alias{ml.train}
\title{Configuration of input file and training data with a machine learning model}
\usage{
ml.train(train_sample, algorithm, metric, seed, resampling)
}
\arguments{
\item{train_sample}{your train set}

\item{algorithm}{specific method from caret::trainControl}

\item{metric}{specific evaluation metric from caret::trainControl}

\item{seed}{specific seed value}

\item{resampling}{specific method from caret::trainControl}
}
\value{
an object is returned of class train
}
\description{
Configuration of input file and training data with a machine learning model
}
\examples{
# load packages
library(mlbench)
library(caret)
library(mlPipe)
# load data
# table available in caret package database
data(PimaIndiansDiabetes)
# rename dataset to keep code below generic
my_df <- PimaIndiansDiabetes
# now we're going to define our training sample
my_training_sample <- ml.trainset(my_df, 0.8, seed = 42)
my_training_sample
# and now we're able to train our model
my_model <- ml.train(train_sample = my_training_sample, algorithm = "lda", metric = "Accuracy", seed = 42, resampling = "cv")
my_model
}
