% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-ml.trainlist.R
\name{ml.trainlist}
\alias{ml.trainlist}
\title{Returning a list from multiple machine learning trained models}
\usage{
ml.trainlist(train_sample, algorithm, metric, seed, resampling, ...)
}
\arguments{
\item{train_sample}{your train set}

\item{algorithm}{set of methods from caret::trainControl}

\item{metric}{specific evaluation metric from caret::trainControl}

\item{seed}{specific seed value}

\item{resampling}{specific method from caret::trainControl}

\item{...}{another parameters from your choice}
}
\value{
a list of objects of class 'train' (from caret::train)
}
\description{
Returning a list from multiple machine learning trained models
}
\examples{
# load packages
library(mlbench)
library(caret)
library(mlPipe)
# load data
# table available in caret package database
data(PimaIndiansDiabetes)
# rename dataset to keep code below generic
my_df <- PimaIndiansDiabetes
# now we're going to define our training sample
my_training_sample <- ml.trainset(my_df, 0.8, seed = 42)
my_training_sample
# and now we're able to train our model
my_models <- ml.trainlist(train_sample = my_training_sample, algorithm = c("rf", "lda", "glm", "knn"), metric = "Accuracy", seed = 42, resampling = "cv")
my_models
}
