% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-ml.trainset.R
\name{ml.trainset}
\alias{ml.trainset}
\title{Setting a training sample to machine learning analyses}
\usage{
ml.trainset(dataset, train_percent, seed)
}
\arguments{
\item{dataset}{a dataframe}

\item{train_percent}{numerical value (percentage, in decimals) you want for training data}

\item{seed}{specific seed value}
}
\value{
a new dataframe, now only with our training sample
}
\description{
Setting a training sample to machine learning analyses
}
\examples{
# load packages
library(mlbench)
library(caret)
library(mlPipe)
# load data
# table available in caret package database
data(PimaIndiansDiabetes)
# rename dataset to keep code below generic
my_df <- PimaIndiansDiabetes
# now you're able to define your training sample
my_training_sample <- ml.trainset(my_df, 0.8, seed = 42)
my_training_sample
}
