% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{mlp_classifier}
\alias{mlp_classifier}
\title{Multilayer perceptron for classification}
\usage{
mlp_classifier(...)
}
\arguments{
\item{...}{initialization arguments for \code{MLP} class}
}
\description{
Multilayer perceptron for classification
}
\examples{
data(iris)
x <- iris[,-5]
y <- iris[,5]
tr <- c(sample(1:50, 25), sample(51:100, 25), sample(101:150, 25))

m <- mlp_classifier(num_epoch=100)
m$fit(x[tr,], y[tr])
table(y[-tr], m$predict(x[-tr,]))
}

