% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{mlp_regressor}
\alias{mlp_regressor}
\title{Multilayer perceptron for regression}
\usage{
mlp_regressor(...)
}
\arguments{
\item{...}{initialization arguments for \code{MLP} class}
}
\description{
Multilayer perceptron for regression
}
\examples{
set.seed(123)
n <- 1000
x <- runif(2*n)
dim(x) <- c(n, 2)
y <- pmin(x[,1], x[,2])
m <- mlp_regressor(hidden_sizes=10, num_epoch=500, batch_size=25)
m$fit(x, y)

newx <- expand.grid(x1=seq(0, 1, length=50), x2=seq(0, 1, length=50))
pred <- m$predict(newx)
true <- pmin(newx[,1], newx[,2])
cor(true, pred)
\dontrun{
dim(pred) <- c(50, 50)
dim(true) <- c(50, 50)
par(mfrow=c(1,2))
contour(true)
contour(pred)}
}

