% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Feature extraction by principal component analysis}
\usage{
pca(ncomp, ...)
}
\arguments{
\item{ncomp}{number of principal components to extract}
}
\description{
Feature extraction by principal component analysis
}
\examples{
p <- pca(2)
p$fit(list(x=USArrests))
p$transform(list(x=USArrests))
p$predict(list(x=USArrests))
}

