% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_extractor}
\alias{pca_extractor}
\title{Feature extraction by principal component analysis}
\usage{
pca_extractor(...)
}
\arguments{
\item{...}{initialization parameters for \code{PCAExtractor} class}
}
\description{
Feature extraction by principal component analysis
}
\examples{
p <- pca_extractor(2, center=TRUE, scale=TRUE)
p$fit(USArrests)
p$transform(USArrests)
p$predict(USArrests)
}

