% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-component.R
\name{pipe_compoenent}
\alias{pipe_compoenent}
\title{Create a custom component for pipeline}
\usage{
pipe_compoenent(...)
}
\arguments{
\item{...}{initialization arguments for \code{PipeComponent} class}
}
\value{
\code{PipeComponent} class object
}
\description{
Create a custom component for pipeline
}

