% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-component.R
\docType{data}
\name{PipeComponent}
\alias{PipeComponent}
\title{Custom component for pipeline}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PipeComponent
}
\description{
Custom component for pipeline
}
\keyword{datasets}

