% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Machine learning pipeline}
\usage{
pipeline(...)
}
\arguments{
\item{...}{Arbitrary number of pipeline components}
}
\description{
Machine learning pipeline
}
\examples{
library(mlbench)
data(Sonar)
X <- Sonar[, -ncol(Sonar)]
y <- Sonar[, ncol(Sonar)]
samp <- c(sample(1:100,75), sample(101:200,75))
m <- mlp_classifier(size=10)
m$fit(list(x=X[samp,], y=class.ind(y[samp])))
table(y[-samp], m$predict(list(x=X[-samp,])))

p <- pipeline(pc=pca(30), ml=mlp_classifier(size=10))
p$fit((list(x=X[samp,], y=class.ind(y[samp]))))
table(y[-samp], p$predict(list(x=X[-samp,])))
}

