% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-component.R
\name{validate_pipe_component_class}
\alias{validate_pipe_component_class}
\title{Validate pipe component class}
\usage{
validate_pipe_component_class(x, verbose = 2)
}
\arguments{
\item{x}{R object, typically an \code{R6ClassGenerator}}

\item{verbose}{integer. if 0, no message is given. if 1, reports only invalidity. it 2+, reports all}
}
\value{
logical that indicates that \code{x} satisfies the minimum conditions for a valid pipe component class
}
\description{
Check if an R6 class satisfies the minimum conditions for pipe components
}
\details{
This function is used by developers of pipe component to check if a class satisfies the minimum properties for a pipe component class.  It is also used by the package developer for package test process to check the validity of the exported pipe component classes.
}
\examples{
validate_pipe_component_class(MLP)
validate_pipe_component_class(PipeComponent)
}
\seealso{
\code{\link{custom_pipe_component}}
}

