% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-ml.train.R
\name{ml.train}
\alias{ml.train}
\title{Configuration of input file and training data with a machine learning model}
\usage{
ml.train(dataset, train_percent, algorithm, metric, seed, resampling)
}
\arguments{
\item{dataset}{a dataframe}

\item{train_percent}{numerical value (percentage, in decimals) you want for training data}

\item{algorithm}{specific method from caret::trainControl}

\item{metric}{specific evaluation metric from caret::trainControl}

\item{seed}{specific seed value}

\item{resampling}{specific method from caret::trainControl}
}
\value{
an object is returned of class train
}
\description{
Configuration of input file and training data with a machine learning model
}
\examples{
# upload the necessary packages
library(mlbench)
library(caret)
library(mlPipe)
# load data
# table available in caret package database
data(PimaIndiansDiabetes)
# rename dataset to keep code below generic
my_df <- PimaIndiansDiabetes
# now you're able to train your data
my_model <- ml.train(my_df, 0.8, "lda", "Accuracy", 42, "repeatedcv")
my_model
}
