% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{mlp}
\alias{mlp}
\title{Multilayer perceptron}
\usage{
mlp(...)
}
\arguments{
\item{...}{initialization arguments for \code{MLP} class}
}
\value{
\code{MLP} class object
}
\description{
Multilayer perceptron
}
\examples{
# example from \\code{\\link{deepnet::nn.train}}
Var1 <- c(rnorm(50, 1, 0.5), rnorm(50, -0.6, 0.2))
Var2 <- c(rnorm(50, -0.8, 0.2), rnorm(50, 2, 1))
x <- matrix(c(Var1, Var2), nrow = 100, ncol = 2)
y <- c(rep(1, 50), rep(0, 50))

m <- mlp(hidden_sizes=5, learn_rate=0.8, num_epoch=3)
m$fit(x, y)
}

