% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{PCAExtractor}
\alias{pca}
\alias{pca_extractor}
\title{Feature extraction by principal component analysis}
\description{
Feature extraction by principal component analysis
}
\section{Usage}{

\preformatted{pca_extractor(ncomp, center = TRUE, scale = TRUE)}
}

\section{Arguments}{

\describe{
\item{\code{ncomp}}{number of principal components to extract}
\item{\code{center}}{either a logical value that indicates whether variables should be                centered to zero-mean, or numeric vector of center values}
\item{\code{scale}}{either a logical value that indicates whether variables should be scaled to unit-variance, or numeric vector of scale values}
}
}

\section{Value}{

\code{PCAExtractor} class object
}

\section{Class Methods}{

\describe{
\item{\preformatted{fit(x, y = NULL)}}{conduct principal component analysis of \code{x}}
\item{\preformatted{transform(x, y = NULL)}}{extract principal components of \code{x}}
\item{\preformatted{predict(x, y = NULL)}}{returns all principal components of \code{x}}
}
}

\section{Details}{

Uses \code{\link[stats]{prcomp}} as the backend.
}
\examples{
p <- pca_extractor(2, center=TRUE, scale=TRUE)
p$fit(USArrests)
p$transform(USArrests)
p$predict(USArrests)
# inv_transform partially recovers the original data
cor(as.numeric(as.matrix(USArrests)),
    as.numeric(p$inv_transform(p$transform(USArrests)$x)$x))

p2 <-  pca_extractor(4, center=TRUE, scale=TRUE)
p2$fit(USArrests)
p2$transform(USArrests)
p2$predict(USArrests)
with ncomp = ncol(x), inv_transform recovers the data perfectly
cor(as.numeric(as.matrix(USArrests)),
    as.numeric(p2$inv_transform(p2$transform(USArrests)$x)$x))
}

