% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_extractor}
\alias{pca_extractor}
\title{Feature extraction by principal component analysis}
\usage{
pca_extractor(...)
}
\arguments{
\item{ncomp}{number of principal components to extract}

\item{center}{a logical indicating whether variables should be shifted to
zero-centered, or numeric vector specifying the arbitrary center values}

\item{scale}{a logical indicating whether variables should be scaled to
unit-variance, or numeric vector specifying the arbitrary scale}
}
\description{
Feature extraction by principal component analysis
}
\examples{
p <- pca_extractor(2, center=TRUE, scale=TRUE)
p$fit(USArrests)
p$transform(USArrests)
p$predict(USArrests)
}

