% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-component.R
\name{pipe_compoenent}
\alias{pipe_compoenent}
\title{Create a custom component for pipeline}
\usage{
pipe_compoenent(fit = NULL, transform = NULL, predict = NULL,
  incfit = NULL, ...)
}
\arguments{
\item{fit}{The function that creates a model object}

\item{predict}{The function that makes prediction from new data}

\item{incfit}{The function that updates the model object}

\item{...}{Parameters to be passed to other methods.}

\item{transofrm}{The function that transforms data.}
}
\value{
\code{PipeComponent} class object
}
\description{
Create a custom component for pipeline
}

