% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-component.R
\name{pipe_component}
\alias{PipeComponent}
\alias{pipe_component}
\title{Default (do-nothing) pipeline component}
\description{
This is a pipeline component doing nothing.
It is mainly used like a virtual class,
where functions are defined in children classes.
}
\section{Usage}{

\preformatted{pipe_component()}
}

\section{Value}{

\code{PipeComponent} class object
}

\section{Class Attributes}{

\describe{
\item{\code{objects}}{an R object. Typically used to store a model object}
}
}

\section{Class Methods}{

\describe{
\item{\preformatted{fit(x=NULL, y=NULL)}}{do nothing}
\item{\preformatted{transpose(x=NULL, y=NULL)}}{return \code{x} and \code{y} as-is}
\item{\preformatted{predict(x=NULL, y=NULL)}}{return \code{NULL}}
\item{\preformatted{incr_fit(x=NULL, y=NULL)}}{do nothing}
\item{\preformatted{predict_proba(x=NULL, y=NULL)}}{return \code{NULL}}
\item{\preformatted{inv_transform(x=NULL, y=NULL)}}{return \code{x} and \code{y} as-is}
}
}
\examples{
p <- pipe_component()
p$fit(x=c(1,5,8), y=c('a','b','a'))  # nothing happens
p$transform(x=1:10, y=1:10)          # return data as-is

p$set_parameters(object=1) # use this to update attributes
# but cannot do so for functions
#p$set_parameters(fit = function() { 'hello world' })
}

