% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Machine learning pipeline}
\usage{
pipeline(...)
}
\arguments{
\item{...}{Arbitrary number of pipeline components}
}
\value{
\code{Pipeline} class object
}
\description{
Machine learning pipeline
}
\examples{
set.seed(123)
data(Sonar, package='mlbench')
X <- Sonar[, -ncol(Sonar)]
y <- Sonar[, ncol(Sonar)]
tr <- c(sample(1:111,75), sample(112:200,75))

p <- pipeline(pc=pca_extractor(30),
              ml=mlp_classifier(hidden_sizes=c(5, 5), num_epoch=1000))
p$fit(X[tr,], y[tr])
table(y[-tr], p$predict(X[-tr,]))
}

